requirements.txt
python-telegram-bot>=12.0.0
SQLAlchemy>=1.1.0
